Allow routing helpers to be used in feature specs
To start off with I was using hardcoded strings for paths but I would
rather use the routing helpers which means that if the route to an
action changes, the specs would break and I won't have to search for
that hardcoded route in my specs.

I needed to create a module to get this to work as just including
Lotus::Helpers::RoutingHelper results in the error RSpec::Routing is
undefined. I've asked the Lotus team if there is a better way to do
this.