Add hardcoded user list and specs
Before dealing with data I wanted to add a page with hardcoded users
and make sure that everything was wired up and rendering correctly.

Turns out that when using Slim you need to use `== yield` rather than
single equals in the layout otherwise the view is escaped when it
doesn't need to be. Inside a template you should still use single
equals when you want to render data from your view as that should be
esacaped by default.